+++ CorePartition/Resources/Thread/Thread.ino
@@ -334,12 +334,12 @@ void Thread2 (void* pValue)
 
 
 
-static uint64_t getTimeTick()
+static uint32_t getTimeTick()
 {
-   return (uint64_t) millis();
+   return (uint32_t) millis();
 }
 
-static void sleepTick (uint64_t nSleepTime)
+static void sleepTick (uint32_t nSleepTime)
 {
     if (nSleepTime)  delayMicroseconds  (nSleepTime * 1000);
 }
